NoNameShadowing: don't report when implicit 'it' parameter isn't used (#3793)
